fix: unify display redactor and address review findings
Review feedback from Codex and Claude (P1-P3):

P1 — The extraction silently replaced the display redactor: old
builtin-preview.ts used packages/ui/src/redact.ts, new tool-quiet-preview.ts
used packages/core/src/redaction.ts. The two are NOT equivalent — the ui
version masks Authorization=Bearer and long hex-with-hyphens tokens that
core's version leaks. Fix: extract ui's redact.ts into core as
display-redaction.ts (single source of truth for display redaction), add
a subpath export, and have both tool-quiet-preview.ts and ui import from it.

P2 — TUI generic input fallback lost the 600-char limit that the old
formatUnknownInline path had. Fix: apply limitText(line, 600) to the
formatToolInvocationLine result before returning.

P3 — English singular '1 replacements' → '1 replacement'.

Closes #1065.
